testing casting in 10

+++ 0x02-functions_nested_loops/10-add.c
@@ -12,6 +12,6 @@ int add(int a, int b)
 {
 	int sum;
 
-	sum = a + b;
+	sum = (int) (a + b);
 	return (sum);
 }
